basic styles system (WIP)
