+++ resources/Themes/megahack.json
@@ -1,0 +1,9 @@
+{
+    "colors": [
+        [0, "ffffffff"],
+        [1, "7d8080ff"],
+        [2, "292929ff"],
+        [7, "212121ff"]
+    ],
+    "floats": []
+}
