+++ src/modules/gui/layouts/window/animation/easing.hpp
