+++ src/modules/gui/layouts/window/animation/move-action.hpp
